rag/requirements.txt
langchain>=0.0.300
langchain-community>=0.0.10
langchain-openai>=0.0.5
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
pypdf>=3.17.4
python-docx>=1.1.0
jieba>=0.42.1
pandas>=2.0.0
numpy>=1.21.0
scikit-learn>=1.3.0
rouge-score>=0.1.2
tqdm>=4.66.0
openai>=1.3.0
transformers>=4.35.0
torch>=2.0.0
datasets>=2.15.0
streamlit>=1.28.0 